gpac_agenda/backend/requirements.txt
# Web Framework
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0

# Environment and Authentication
python-dotenv==0.19.0
pyjwt==2.1.0
bcrypt==3.2.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Production Dependencies
werkzeug==2.0.1
